Add a flag to reindex "current" indices
Index with a name matching the current date are not reindexed because we
want to reindex all that can safely be without interupting ElasticSearch
documents ingestion.

When these non-current indexes are reindexed, we can stop the services
that inject data to ElasticSearch, reindex *all* indices, replace
ElasticSearch with OpenSearch and restart the services that produce
data.

While here, remove unused variable.