Fix detection of existing indices
We want an exact match to determine if a given was reindexed or not.  If
a daily index `.samplerr-2022.11.21` was imported in a previous run,
monthly index `.samplerr-2022.11` and yearly index `.samplerr-2022`
should not be skipped just because they match `.samplerr-2022.11.21`.

Add boundary checks to the match regular expression to fix this.